Dev/alex

+++ application/conditions/controller.py
@@ -4,14 +4,14 @@ from .model import Condition
 from .. import db
 import base64
 
-from flask_jwt_extended import (
-    create_access_token,
-    create_refresh_token,
-    jwt_required,
-    get_jwt,
-    current_user,
-    get_jwt_identity,
-)
+# from flask_jwt_extended import (
+#     create_access_token,
+#     create_refresh_token,
+#     jwt_required,
+#     get_jwt,
+#     current_user,
+#     get_jwt_identity,
+# )
 
 from flask_jwt_extended import jwt_required, get_jwt
 from application.patients.model import Patient, TokenBlocklist
@@ -19,17 +19,20 @@ from schemas import UserSchema
 
 
 def get_conditions():
-    # claims = get_jwt()
+    claims = get_jwt()
 
-    # if claims.get("is_staff") == True:
+    if claims.get("is_staff") == True:
         condition = Condition.query.all()
         try:
             return jsonify({ "data": [c.json for c in condition] }), 200
         except:
             raise exceptions.InternalServerError(f"Conditions not found!")
+        
+    return jsonify({"message": "You are not authorized to access this"}), 401
     
 
 def get_user_conditions(patient_email):
+
     print("patient_id", type(id))
     conditions = Condition.query.filter_by(patient_email=patient_email).all()
     try:
@@ -63,18 +66,30 @@ def create_condition():
         raise exceptions.BadRequest(f"We cannot process your request")
 
 def update_condition(id):
-    data = request.json
-    condition = Condition.query.filter_by(id=id).first()
+    claims = get_jwt()
 
-    for (attribute, value) in data.items():
-        if hasattr(condition, attribute):
-            setattr(condition, attribute, value)
+    if claims.get("is_staff") == True:
+        data = request.json
+        condition = Condition.query.filter_by(id=id).first()
 
-    db.session.commit()
-    return jsonify({ "data": condition.json })
+        for (attribute, value) in data.items():
+            if hasattr(condition, attribute):
+                setattr(condition, attribute, value)
 
-def destroy_condition(id):
-    condition = Condition.query.filter_by(id=id).first()
-    db.session.delete(condition)
-    db.session.commit()
-    return "Condition Deleted", 204
+        db.session.commit()
+        return jsonify({ "data": condition.json })
+    
+    return jsonify({"message": "You are not authorized to access this"}), 401
+
+
+def destroy_condition(id): 
+    claims = get_jwt()
+
+    if claims.get("is_staff") == True:
+
+        condition = Condition.query.filter_by(id=id).first()
+        db.session.delete(condition)
+        db.session.commit()
+        return "Condition Deleted", 204
+
+    return jsonify({"message": "You are not authorized to access this"}), 401
